Refactored TokenURISet event test
